Helmfile: Install on demand

+++ charts/scripts/helmfile.sh
@@ -18,9 +18,13 @@ HELM_DIFF_VERSION="3.1.3"
 mkdir -p /usr/lib/google-cloud-sdk/bin
 ln -s /builder/google-cloud-sdk/bin/gcloud /usr/lib/google-cloud-sdk/bin/gcloud
 
-echo "Installing Helm Diff..."
-helm plugin install https://github.com/databus23/helm-diff --version "${HELM_DIFF_VERSION}" \
-  >>/dev/null
+if helm plugin list | grep -E '^diff ' --quiet; then
+  echo "Helm Diff is already installed (presumably because ${BASH_SOURCE[0]} was run earlier)"
+else
+  echo "Installing Helm Diff..."
+  helm plugin install https://github.com/databus23/helm-diff --version "${HELM_DIFF_VERSION}" \
+    >>/dev/null
+fi
 
 helmfile "$@"
